requirements.txt.txt
streamlit
pandas
vnstock
pandas-ta
plotly
google-generativeai